refactor: rating 도메인 전용 localStorage 로직을 ratingStorage로 분리 및 이동
