+++ src/domain/rating/constants/storageKeys.js
