+++ src/domain/rating/ratingStorage.js
